GUI Popups
This version includes a .sh shell script that should work on most *nix operating systems, as well as a .command file that can be executed on mac by double click. If the file does not have execute permissions set, then it may not work by double click. Run chmod +x (path to .command file) to add execute permission and run on double click. This script will provide GUI prompts for input, and output an HTML file to the user's desktop. This has yet to be wrapped in an app and is a little rough around the edges